Diff conflate bug fix/feature
* Fixes bug in Diff conflate where reference elements were being removed by tag rather than status
* Adds Diff conflate option, `differential.remove.reference.snapped.data`, to also delete reference data involved in snapping in scenarios where changeset output won't be written

+++ hoot-core/src/main/cpp/hoot/core/validation/Validator.h
@@ -59,12 +59,12 @@ public:
   /**
     * @brief Returns the total number of validation errors encountered
     */
-  virtual int getNumValidationErrors() const = 0;
+  virtual long getNumValidationErrors() const = 0;
 
   /**
     * @brief Returns the total number of features validated
     */
-  virtual int getNumFeaturesValidated() const = 0;
+  virtual long getNumFeaturesValidated() const = 0;
 };
 
 }
